Modularised the code with edit/done button.
Updated all profiles (staff and patient) to use the new button & code.
Edited button code to allow updating of the database.
