+++ components/profile/patient/AboutTab.js
@@ -2,42 +2,15 @@ import React from "react";
 import { useState } from "react";
 import { StyleSheet, View, Text, TouchableOpacity } from "react-native";
 import ProfileInformation from "../profilecomponents/ProfileInformation";
-import { MaterialIcons } from "@expo/vector-icons";
 import colorDefaults from "../../../theme/colorDefaults";
+import EditEnableButton from "../profilecomponents/EditEnableButton";
 
 const PatientAboutTab = () => {
   const [firstName, setFirstName] = useState("John");
   const [middleName, setMiddleName] = useState("");
   const [lastName, setLastName] = useState("Doe");
-  const [enabled, setEnabled] = useState(false);
+  const [editable, setEditable] = useState(false);
 
-  function editText() {
-    enabled ? setEnabled(false) : setEnabled(true);
-  }
-
-  const doneIcon = (
-    <View style={styles.icon}>
-      <MaterialIcons
-        name="done"
-        size={24}
-        color="black"
-        style={{ marginRight: 5 }}
-      />
-      <Text>Done</Text>
-    </View>
-  );
-
-  const editIcon = (
-    <View style={styles.icon}>
-      <MaterialIcons
-        name="edit"
-        size={24}
-        color="black"
-        style={{ marginRight: 5 }}
-      />
-      <Text>Edit</Text>
-    </View>
-  );
   return (
     <View style={styles.profiles}>
       <ProfileInformation
@@ -45,25 +18,24 @@ const PatientAboutTab = () => {
         placeholder="First Name"
         value={firstName}
         onChangeText={setFirstName}
-        editable={enabled}
+        editable={editable}
       />
       <ProfileInformation
         label="Middle Name"
         placeholder="Middle Name"
         value={middleName}
         onChangeText={setMiddleName}
-        editable={enabled}
+        editable={editable}
       />
       <ProfileInformation
         label="Last Name"
         placeholder="Last Name"
         value={lastName}
         onChangeText={setLastName}
-        editable={enabled}
+        editable={editable}
       />
-      <TouchableOpacity style={styles.button} onPress={editText}>
-        <Text>{enabled ? doneIcon : editIcon}</Text>
-      </TouchableOpacity>
+
+      <EditEnableButton editable={editable} setEditable={setEditable} />
     </View>
   );
 };
