+++ components/profile/patient/MedicalTab.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { useState } from "react";
 import {
   Keyboard,
   KeyboardAvoidingView,
